Fix for testBug72630.

+++ src/testsuite/regression/CharsetRegressionTest.java
@@ -86,36 +86,42 @@ public class CharsetRegressionTest extends BaseTestCase {
      * @throws Exception
      */
     public void testBug72630() throws Exception {
-        try {
-            this.stmt.execute("CREATE USER 'Bug72630User'@'%' IDENTIFIED WITH mysql_native_password AS 'pwd'");
-            this.stmt.execute("GRANT ALL ON *.* TO 'Bug72630User'@'%'");
-
-            final Properties props = new Properties();
-            props.setProperty("user", "Bug72630User");
-            props.setProperty("password", "pwd");
-            props.setProperty("characterEncoding", "NonexistentEncoding");
-
-            assertThrows(SQLException.class, "Unsupported character encoding 'NonexistentEncoding'.", new Callable<Void>() {
-                public Void call() throws Exception {
-                    getConnectionWithProps(props);
-                    return null;
+        // bug is related to authentication plugins, available only in 5.5.7+ 
+        if (versionMeetsMinimum(5, 5, 7)) {
+            try {
+                if (versionMeetsMinimum(5, 5)) {
+                    this.stmt.execute("CREATE USER 'Bug72630User'@'%' IDENTIFIED WITH mysql_native_password AS 'pwd'");
+                } else {
+                    this.stmt.execute("CREATE USER 'Bug72630User'@'%' IDENTIFIED BY 'pwd'");
                 }
-            });
+                this.stmt.execute("GRANT ALL ON *.* TO 'Bug72630User'@'%'");
 
-            props.remove("characterEncoding");
-            props.setProperty("passwordCharacterEncoding", "NonexistentEncoding");
-            assertThrows(SQLException.class, "Unsupported character encoding 'NonexistentEncoding' for 'passwordCharacterEncoding' or 'characterEncoding'.",
-                    new Callable<Void>() {
-                        public Void call() throws Exception {
-                            getConnectionWithProps(props);
-                            return null;
-                        }
-                    });
-        } catch (SQLException e) {
-            e.printStackTrace();
-        } finally {
-            this.stmt.execute("DROP USER 'Bug72630User'@'%'");
+                final Properties props = new Properties();
+                props.setProperty("user", "Bug72630User");
+                props.setProperty("password", "pwd");
+                props.setProperty("characterEncoding", "NonexistentEncoding");
+
+                assertThrows(SQLException.class, "Unsupported character encoding 'NonexistentEncoding'.", new Callable<Void>() {
+                    public Void call() throws Exception {
+                        getConnectionWithProps(props);
+                        return null;
+                    }
+                });
+
+                props.remove("characterEncoding");
+                props.setProperty("passwordCharacterEncoding", "NonexistentEncoding");
+                assertThrows(SQLException.class,
+                        "Unsupported character encoding 'NonexistentEncoding' for 'passwordCharacterEncoding' or 'characterEncoding'.", new Callable<Void>() {
+                            public Void call() throws Exception {
+                                getConnectionWithProps(props);
+                                return null;
+                            }
+                        });
+            } catch (SQLException e) {
+                e.printStackTrace();
+            } finally {
+                this.stmt.execute("DROP USER 'Bug72630User'@'%'");
+            }
         }
-
     }
 }
